Add uint8 affinity support (Option B: template-parameterized AffValue)
- Move AffValue/ScoreValue typedefs from frontend_agglomerate.h into
  JIT-generated AffType.h — each (dtype, scoring_function) combo gets
  its own cached compiled module
- Add uint8 specializations:
  - discretize.hpp: identity for uint8 with 256 bins (no multiply)
  - Operators.hpp: OneMinus<uint8> = 255 - x, Invert<uint8> = 255/x
- Refactor _agglomerate.py: shared _compile_module() helper, detects
  affs.dtype and patches .pyx source + headers accordingly
- ScoredEdge/Merge structs use double for Python interop
- Fix cout printing for uint8 ScoreType (cast to double)

uint8 mode: 4x less memory for affinities, ~2x better cache utilization
in the voxel scan. Lossless for HistogramQuantileAffinity with 256 bins.

+++ src/waterz/backend/IterativeRegionMerging.hpp
@@ -55,10 +55,10 @@ public:
 				scoreEdge(e, edgeScoringFunction);
 		}
 
-		std::cout << "merging until " << threshold << std::endl;
+		std::cout << "merging until " << (double)threshold << std::endl;
 
 		if (!_edgeQueue.empty())
-			std::cout << "min edge score " << _edgeScores[_edgeQueue.top()] << std::endl;
+			std::cout << "min edge score " << (double)_edgeScores[_edgeQueue.top()] << std::endl;
 
 		// while there are still unhandled edges
 		std::size_t merged = 0;
